In Class stuff done
Missile is firing.
Need to:
Fix trajectory of missile when fired from a non-zero HPR.
Fix extra models being loaded in for everything.
Fix collisions on ''some'' of the drones; the ones around the planets work, but not the ones around the spacestation. May be linked to duplicate model problem mentioned before.